+++ packages/ai/src/ui/ui-messages.ts
@@ -68,7 +68,7 @@ export type DataUIPart<DATA_TYPES extends UIDataTypes> = ValueOf<{
   [NAME in keyof DATA_TYPES & string]: {
     type: `data-${NAME}`;
     id?: string;
-    value: DATA_TYPES[NAME];
+    data: DATA_TYPES[NAME];
   };
 }>;