+++ packages/react/src/use-chat.ui.test.tsx
@@ -30,16 +30,16 @@ describe('data protocol stream', () => {
   let onFinishCalls: Array<{ message: UIMessage }> = [];
 
   setupTestComponent(
-    ({ id: idParam }: { id: string }) => {
-      const [id, setId] = React.useState<string>(idParam);
+    ({ chatId: idParam }: { chatId: string }) => {
+      const [chatId, setChatId] = React.useState<string>(idParam);
       const {
         messages,
         append,
         error,
         status,
-        id: idKey,
+        chatId: idKey,
       } = useChat({
-        id,
+        chatId,
         onFinish: options => {
           onFinishCalls.push(options);
         },
@@ -64,7 +64,7 @@ describe('data protocol stream', () => {
           <button
             data-testid="do-change-id"
             onClick={() => {
-              setId('second-id');
+              setChatId('second-id');
             }}
           />
         </div>
@@ -313,7 +313,7 @@ describe('data protocol stream', () => {
 
       expect(await server.calls[0].requestBodyJson).toMatchInlineSnapshot(`
         {
-          "id": "first-id-0",
+          "chatId": "id-0",
           "messages": [
             {
               "id": "id-1",
@@ -635,7 +635,7 @@ describe('prepareRequestBody', () => {
 
     expect(bodyOptions).toMatchInlineSnapshot(`
       {
-        "id": "id-0",
+        "chatId": "id-0",
         "messages": [
           {
             "id": "msg-0",
@@ -1407,7 +1407,7 @@ describe('file attachments with data url', () => {
 
     expect(await server.calls[0].requestBodyJson).toMatchInlineSnapshot(`
       {
-        "id": "id-0",
+        "chatId": "id-0",
         "messages": [
           {
             "id": "id-0",
@@ -1490,7 +1490,7 @@ describe('file attachments with data url', () => {
 
     expect(await server.calls[0].requestBodyJson).toMatchInlineSnapshot(`
       {
-        "id": "id-0",
+        "chatId": "id-0",
         "messages": [
           {
             "id": "id-0",
@@ -1605,7 +1605,7 @@ describe('file attachments with url', () => {
 
     expect(await server.calls[0].requestBodyJson).toMatchInlineSnapshot(`
       {
-        "id": "id-0",
+        "chatId": "id-0",
         "messages": [
           {
             "id": "id-0",
@@ -1711,7 +1711,7 @@ describe('attachments with empty submit', () => {
 
     expect(await server.calls[0].requestBodyJson).toMatchInlineSnapshot(`
       {
-        "id": "id-0",
+        "chatId": "id-0",
         "messages": [
           {
             "id": "id-1",
@@ -1823,7 +1823,7 @@ describe('should append message with attachments', () => {
 
     expect(await server.calls[0].requestBodyJson).toMatchInlineSnapshot(`
       {
-        "id": "id-0",
+        "chatId": "id-0",
         "messages": [
           {
             "id": "id-1",
@@ -1910,7 +1910,7 @@ describe('reload', () => {
 
     expect(await server.calls[1].requestBodyJson).toMatchInlineSnapshot(`
       {
-        "id": "id-0",
+        "chatId": "id-0",
         "messages": [
           {
             "id": "id-1",
@@ -1983,7 +1983,7 @@ describe('test sending additional fields during message submission', () => {
 
     expect(await server.calls[0].requestBodyJson).toMatchInlineSnapshot(`
       {
-        "id": "id-0",
+        "chatId": "id-0",
         "messages": [
           {
             "id": "id-1",
@@ -2023,7 +2023,7 @@ describe('resume ongoing stream and return assistant message', () => {
       ]);
 
       const { messages, status, experimental_resume } = useChat({
-        id: '123',
+        chatId: '123',
         chatStore,
       });