+++ packages/langchain/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @ai-sdk/langchain
 
+## 1.0.0-alpha.1
+
+### Patch Changes
+
+- Updated dependencies [b346545]
+- Updated dependencies [109c0ac]
+  - ai@5.0.0-alpha.1
+
 ## 1.0.0-canary.12
 
 ### Patch Changes